Style homepage grid
By using enumerate and zip within the homepage queryset, the categories and posts can be combined in a way that the template can build sets of one each. This can then be worked within the css grid to get the column and row flow I have in mind. The post and categories will always be limited to 6 on the homepage but their dedicated pages will show the remainder.